Routex.Extension behaviour (Routex v1.1.0)

Specification for composable Routex extensions.

All callbacks are optional

See also: Routex Extensions

Types

attrs()

@type attrs() :: map()

backend()

@type backend() :: Routex.Backend.t()

conn()

@type conn() :: Plug.Conn.t()

env()

@type env() :: Macro.Env.t()

opts()

@type opts() :: list()

params()

@type params() :: map()

routes()

@type routes() :: [Phoenix.Router.Route.t()]

socket()

@type socket() :: Phoenix.Socket.t()

url()

@type url() :: binary()

Callbacks

after_render(params, url, socket, attrs)
(optional)

@callback after_render(params(), url(), socket(), attrs()) ::
  {:cont,
   %Phoenix.Socket{
     assigns: term(),
     channel: term(),
     channel_pid: term(),
     endpoint: term(),
     handler: term(),
     id: term(),
     join_ref: term(),
     joined: term(),
     private: term(),
     pubsub_server: term(),
     ref: term(),
     serializer: term(),
     topic: term(),
     transport: term(),
     transport_pid: term()
   }}

Callback for the after_render livecycle stage with the same name. Receives an additional argument attrs with Routex.Attrs of the current route. See also Livecycle callbacks

The callback should return either {:cont, socket} or {:halt, socket}

configure(opts, backend)
(optional)

@callback configure(opts(), backend()) :: opts()

The configure/2 callback is called in the first stage with the options provided to Routex and the name of the Routex backend. It is expected to return a new list of options.

create_helpers(routes, backend, env)
(optional)

@callback create_helpers(routes(), backend(), env()) :: Macro.output()

The create_helpers/3 callback is called in the last stage with a list of routes belonging to a Routex backend, the name of the Routex backend and the current environment. It is expected to return Elixir AST.

The AST is included in MyAppWeb.Router.RoutexHelpers.

handle_async(params, url, socket, attrs)
(optional)

@callback handle_async(params(), url(), socket(), attrs()) ::
  {:cont,
   %Phoenix.Socket{
     assigns: term(),
     channel: term(),
     channel_pid: term(),
     endpoint: term(),
     handler: term(),
     id: term(),
     join_ref: term(),
     joined: term(),
     private: term(),
     pubsub_server: term(),
     ref: term(),
     serializer: term(),
     topic: term(),
     transport: term(),
     transport_pid: term()
   }}

Callback for the handle_async livecycle stage with the same name. Receives an additional argument attrs with Routex.Attrs of the current route. See also Livecycle callbacks

The callback should return either {:cont, socket} or {:halt, socket}

handle_event(params, url, socket, attrs)
(optional)

@callback handle_event(params(), url(), socket(), attrs()) ::
  {:cont,
   %Phoenix.Socket{
     assigns: term(),
     channel: term(),
     channel_pid: term(),
     endpoint: term(),
     handler: term(),
     id: term(),
     join_ref: term(),
     joined: term(),
     private: term(),
     pubsub_server: term(),
     ref: term(),
     serializer: term(),
     topic: term(),
     transport: term(),
     transport_pid: term()
   }}

Callback for the handle_event livecycle stage with the same name. Receives an additional argument attrs with Routex.Attrs of the current route. See also Livecycle callbacks

The callback should return either {:cont, socket} or {:halt, socket}

handle_info(params, url, socket, attrs)
(optional)

@callback handle_info(params(), url(), socket(), attrs()) ::
  {:cont,
   %Phoenix.Socket{
     assigns: term(),
     channel: term(),
     channel_pid: term(),
     endpoint: term(),
     handler: term(),
     id: term(),
     join_ref: term(),
     joined: term(),
     private: term(),
     pubsub_server: term(),
     ref: term(),
     serializer: term(),
     topic: term(),
     transport: term(),
     transport_pid: term()
   }}

Callback for the handle_info livecycle stage with the same name. Receives an additional argument attrs with Routex.Attrs of the current route. See also Livecycle callbacks

The callback should return either {:cont, socket} or {:halt, socket}

handle_params(params, url, socket, attrs)
(optional)

@callback handle_params(params(), url(), socket(), attrs()) ::
  {:cont,
   %Phoenix.Socket{
     assigns: term(),
     channel: term(),
     channel_pid: term(),
     endpoint: term(),
     handler: term(),
     id: term(),
     join_ref: term(),
     joined: term(),
     private: term(),
     pubsub_server: term(),
     ref: term(),
     serializer: term(),
     topic: term(),
     transport: term(),
     transport_pid: term()
   }}

Callback for the handle_params livecycle stage with the same name. Receives an additional argument attrs with Routex.Attrs of the current route. See also Livecycle callbacks

The callback should return either {:cont, socket} or {:halt, socket}

plug(conn, opts, attrs)
(optional)

@callback plug(conn(), opts(), attrs()) :: conn()

Callback for the Plug pipeline. Receives an additional argument attrs with Routex.Attrs of the current route. The callback should return conn

post_transform(routes, backend, env)
(optional)

@callback post_transform(routes(), backend(), env()) :: routes()

The post_transform/1 callback is called in the third stage with a list of routes belonging to a Routex backend. It is expected to return a list of Phoenix.Router.Route structs almost identical to the input, only adding Routex.Attrs -for own usage- is allowed.

transform(routes, backend, env)
(optional)

@callback transform(routes(), backend(), env()) :: routes()

The transform/3 callback is called in the second stage with a list of routes belonging to a Routex backend, the name of the configuration model and the current environment. It is expected to return a list of Phoenix.Router.Route structs with flattened Routex.Attrs.